Issue #148: Use CoroutineScopes in NavigationOverlayFragment's views.
I removed the UNDISPATCHED code because it's a micro-optimization and
uses experimental APIs.

I tested this worked by adding a delay to the code:

    uiScope.launch {
        delay(3000)
        Log.e("lol", "I am called")
        // Bind home tile...

I opened the overlay, dismissed the overlay, and read the logcat. With
the cancellation code (the final implementation), the statement was not
printeed. Without the cancellation code (commented out), the statement
was printed.