Issue #148: Use ActivityCoroutineScope in MainActivity.
I removed the UNDISPATCHED code because it's a micro-optimization and
uses experimental APIs.

I tested the cancellation works correctly by adding the following code:

hideProgressBarJob = uiScope.launch {
    delay(3000)
    Log.e("lol", "I am called")
    progressBar.progress = MIN_UI_PROGRESS
}

I opened the app, started the progress bar, closed the app (with the
back button), and checked logcats. When the cancellation code was there
(the final implementation), the message was not printed. When I
commented out the cancellation code, the message was printed.